Add 'toggle-drag' zoom strategy
This zoom strategy allows users to zoom in on an image and move their mouse to
find the desired area of the image where they want to remain zoomed in. After
finding this area, they can click to hold the zoom in the current position,
allowing them to perform other tasks on the page while keeping zoomed in on the
desired location.

* First click: initiate zoom, follow mouse movements
* Second click: keep zoom, stop following mouse
* Third click: stop zoom (back to initial state)